+++ Net_Core/Web.Angular/projects/gw-frontend/src/app/skins/dashboard/dashboard-header/dashboard-header.component.ts
@@ -1,7 +1,11 @@
 import { Component, computed, input } from '@angular/core';
-import { Router } from '@angular/router';
+import { Router, RouterLink } from '@angular/router';
 // Angular Material
+import { MatButtonModule } from '@angular/material/button';
 import { MatSidenav } from '@angular/material/sidenav';
+import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatToolbarModule } from '@angular/material/toolbar';
 // Library
 import { AccountService } from '@growthware/core/account';
 import { GWCommon } from '@growthware/common/services';
@@ -11,8 +15,17 @@ import { ConfigurationService } from '@growthware/core/configuration';
 
 @Component({
 	selector: 'gw-frontend-dashboard-header',
+	standalone: true,
 	templateUrl: './dashboard-header.component.html',
-	styleUrls: ['./dashboard-header.component.scss']
+	styleUrls: ['./dashboard-header.component.scss'],
+	imports: [
+		RouterLink,
+		
+		MatButtonModule,
+		MatIconModule,
+		MatMenuModule,
+		MatToolbarModule,
+	],
 })
 export class DashboardHeaderComponent {
 
@@ -27,7 +40,8 @@ export class DashboardHeaderComponent {
 		private _ConfigurationSvc: ConfigurationService,
 		private _GWCommon: GWCommon,
 		private _ModalSvc: ModalService,
-		private _Router: Router) { }
+		private _Router: Router
+	) { }
 
 	onLogin(): void {
 		const mWindowSize: WindowSize = new WindowSize(225, 450);