+++ Release_Notes.md
@@ -5,6 +5,7 @@
 - Added Feedback support
 - Converted the Arc "skin" to standalone
 - Converted the Blue Arrow "skin" to standalone
+- Converted the Dashboard "skin" to standalone
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure